Normalize display names and go back to string field types
The normalization of names is done because there are situations like a
column called `Number of Employees` vs `Number Of Employees` (note the
casing difference). This raised an error in the bigquery target and
generally seems like a bad idea.

I went back to strings for the types because shared fields currently do
not expose the pg_data_type key, meaning there's no way to distinguish
integers vs decimal numbers for partner records. Typing can be added
back once this initial splitting of tables is done.