Singer merge (#2)
* Flatten reports. Detect duplicate column names

* Switch to using the new partner-records route

* Cleanup names of stuff for partner records switch

* Schema discovery. Pagination

* Allow using local environment, up partner-records limit to 1k

* loosen project dependencies

* Refactor partner records sync, add support for populations and new crossbeam IDs

* make crossbeam id part of pk

* Implement syncing of new /records route

* do not sync fields if they are not found in the schema

* Revert "do not sync fields if they are not found in the schema"

This reverts commit c6a3c24f507f41c069ac502477bcf4e8ceb00494.

* Working records sync

* Fix records schema and stream selection

* Partner records working

* Refactor discover code

* Refactors and fixes to syncing of records

* Remove the is_partner_data metadata

This was confusing me due to it being used for your own data,
not your partners'. This change clears up which streams we
will actually be syncing and how the metadata for them is
collected.

* implement discovery for partner records

* Change partner records syncing to use separate tables

* Normalize display names and go back to string field types

The normalization of names is done because there are situations like a
column called `Number of Employees` vs `Number Of Employees` (note the
casing difference). This raised an error in the bigquery target and
generally seems like a bad idea.

I went back to strings for the types because shared fields currently do
not expose the pg_data_type key, meaning there's no way to distinguish
integers vs decimal numbers for partner records. Typing can be added
back once this initial splitting of tables is done.

* Implement handling of owner records

* Fix some bugs with syncing owners

* Add back typing, falling back to string types when conflicts

* Ignore sources where mdm_type is missing

This will only happen for deprecated hubspot sources
that we don't want to sync.

* Fix bugs writing partner_user to user and wrong join key

* Set inclusion: automatic on built in record fields

* Set x-requested-with header

* Go back to using mdm_type to look up streams

Using source_id is faulty because there may be sources with no fields
selected, but we still want to include those rows in the result set. So
the mdm_type is needed to be able to find out which stream the things
belong to.

Co-authored-by: Andrew Madonna <andrew.w.madonna@gmail.com>
Co-authored-by: zachharris1 <69470481+zachharris1@users.noreply.github.com>